Model.py added with database updated commit
